// packages/media/transcoder/src/types.ts
export type FFMpegCodecType = 'video' | 'audio' | 'subtitle' | 'data' | 'attachment';
// ...
export interface FFMpegCodecCapability {
  name: string;
  longName: string;
  codecType: FFMpegCodecType;
  supportsDecoding: boolean;
  supportsEncoding: boolean;
  intraFrameOnly: boolean;
  lossyCompression: boolean;
  losslessCompression: boolean;
}